vectorai-service/requirements.txt
fastapi==0.115.0
uvicorn==0.30.0
google-generativeai>=0.8.0
actiancortex-0.1.0b1-py3-none-any.whl
